Create lib directory and refactor code so that generic modules are separate to project modules.
