+++ lib/base_page.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:kar_kam/app_settings.dart';
 import 'package:kar_kam/button_array.dart';
 // import 'package:kar_kam/data_notification.dart';
-import 'package:kar_kam/data_notifier.dart';
+import 'package:kar_kam/lib/data_notifier.dart';
 // import 'package:kar_kam/notification_notifier.dart';
 import 'package:kar_kam/page_specs.dart';
